# openapi_fetched.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kristo Intelligence API",
    "version": "1.0.0",
    "description": "Production-ready pay-per-call AI Agent Intelligence API powered by x402 on Base Blockchain (Chain ID 8453). Each paid endpoint requires payment of 0.01 USDC sent to 0xd4cdA980839C8FED4374EE37EA8DBE8c4ECfd88f via X-PAYMENT header.",
    "x-402-pricing": {
      "price_per_call_usdc": 0.01,
      "amount_raw": "10000",
      "asset": "USDC",
      "asset_address": "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913",
      "pay_to": "0xd4cdA980839C8FED4374EE37EA8DBE8c4ECfd88f",
      "chain_id": 8453
    }
  },
  "paths": {
...
    "/crypto/token-prices": {
      "get": {
        "summary": "Token Prices (Paid)",
        "description": "Requires 0.01 USDC via x402. Returns real-time token prices from CoinGecko for specified Base tokens.",
        "parameters": [
          {
            "name": "tokens",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "ETH,USDC,WETH,BASE"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Real-time token prices"
          },
          "402": {
            "description": "Payment Required (x402)"
          }
        }
      }
    },
...
    "/crypto/gas-oracle": {
      "get": {
        "summary": "Base Gas Oracle (Paid)",
        "description": "Requires 0.01 USDC via x402. Returns current Base gas price, cost estimates for ETH/USDC transfers, swaps, contract calls, and gas recommendation.",
        "responses": {
          "200": {
            "description": "Gas oracle metrics and estimates"
          },
          "402": {
            "description": "Payment Required (x402)"
          }
        }
      }
    }
  }
}